epc-ew-cli/pyproject.toml
[project]
name = "epc-ew-cli"
version = "0.1.0"
description = "CLI for epc-ew (EPC England & Wales consumer)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "epc-ew",
  "typer>=0.12.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
epc-ew = "epc_ew_cli:run"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/epc_ew_cli"]

[tool.uv.sources]
epc-ew = { workspace = true }
